Documentation and refactoring
