+++ app/src/main/java/API/Database/OnDataReceivedFromDB.java
@@ -1,4 +1,4 @@
-package API;
+package API.Database;
 
 import javax.annotation.Nullable;
