+++ app/src/main/java/API/Database/OnDataSentToDB.java
@@ -1,4 +1,4 @@
-package API;
+package API.Database;
 
 import androidx.annotation.NonNull;
