+++ app/src/main/java/UI/BranchViewActivity.java
@@ -16,8 +16,8 @@ import javax.annotation.Nullable;
 import BusinessEntities.Branch;
 import BusinessEntities.Menu;
 import BusinessEntities.QRCode;
-import API.Database;
-import API.OnDataReceivedFromDB;
+import API.Database.Database;
+import API.Database.OnDataReceivedFromDB;
 import DataAccessLayer.RestDB;
 import UIAdapters.MenuRecyclerViewAdapter;
 import ViewModels.MenuViewModel;
